Moved Game class to own file
